+++ docs/home/editor/save_alternative.php
@@ -21,7 +21,7 @@
  */
 
 define('TR_INCLUDE_PATH', '../../include/');
-require (TR_INCLUDE_PATH.'vitals.inc.php');
+require_once(TR_INCLUDE_PATH.'vitals.inc.php');
 
 $pid = intval($_POST['pid']);
 $type_id = intval($_POST['a_type']);
@@ -30,7 +30,9 @@ $secondary_resource = trim($_POST['alternative']);
 // check post vars
 if ($pid == 0 || $type_id == 0 || $secondary_resource == '') exit;
 
-global $db;
+require_once(TR_INCLUDE_PATH.'classes/DAO/DAO.class.php');
+$dao = new DAO();
+
 // delete the existing alternative for this (pid, a_type)
 $sql = "SELECT sr.secondary_resource_id 
           FROM ".TABLE_PREFIX."secondary_resources sr, ".TABLE_PREFIX."secondary_resources_types srt
@@ -38,29 +40,32 @@ $sql = "SELECT sr.secondary_resource_id
            AND sr.primary_resource_id = ".$pid."
            AND sr.language_code = '".$_SESSION['lang']."'
            AND srt.type_id=".$type_id;
-$existing_secondary_result = mysql_query($sql, $db);
+//$existing_secondary_result = mysql_query($sql, $db);
+$existing_secondary_rows = $dao->execute($sql);
 
-while ($existing_secondary = mysql_fetch_assoc($existing_secondary_result))
-{
-	$sql = "DELETE FROM ".TABLE_PREFIX."secondary_resources 
-	         WHERE secondary_resource_id = ".$existing_secondary['secondary_resource_id'];
-	$result = mysql_query($sql, $db);
-
-	$sql = "DELETE FROM ".TABLE_PREFIX."secondary_resources_types 
-	         WHERE secondary_resource_id = ".$existing_secondary['secondary_resource_id']."
-	           AND type_id=".$type_id;
-	$result = mysql_query($sql, $db);
+if (is_array($existing_secondary_rows)) {
+	foreach ($existing_secondary_rows as $existing_secondary)
+	{
+		$sql = "DELETE FROM ".TABLE_PREFIX."secondary_resources 
+		         WHERE secondary_resource_id = ".$existing_secondary['secondary_resource_id'];
+		$dao->execute($sql);
+	
+		$sql = "DELETE FROM ".TABLE_PREFIX."secondary_resources_types 
+		         WHERE secondary_resource_id = ".$existing_secondary['secondary_resource_id']."
+		           AND type_id=".$type_id;
+		$dao->execute($sql);
+	}
 }
 
 // insert new alternative
 $sql = "INSERT INTO ".TABLE_PREFIX."secondary_resources (primary_resource_id, secondary_resource, language_code)
         VALUES (".$pid.", '".mysql_real_escape_string($secondary_resource)."', '".$_SESSION['lang']."')";
-$result = mysql_query($sql, $db);
+$dao->execute($sql);
 $secondary_resource_id = mysql_insert_id();
 
 $sql = "INSERT INTO ".TABLE_PREFIX."secondary_resources_types (secondary_resource_id, type_id)
         VALUES (".$secondary_resource_id.", ".$type_id.")";
-$result = mysql_query($sql, $db);
+$dao->execute($sql);
 
 exit;
